Make PHONEME_LIST2 and related arrays dynamic
This fixes the #845/#945 memory overwrite issue and is a partial step to
a full MT capable version (i.e. a version with no global memory).  The
arrays are extended incrementally via realloc, the algorithm is crude
(to the point of nonexistence), the arrays start off a fixed length and
get stepped up each time by a fixed increment, however because the
arrays are retained across calls using the same Translator there is some
hope that they will be adaptive within a single language stream -
LoadDictionary destroys the Translator so the arrays are reset by a
dictionary change.

+++ src/libespeak-ng/phonemelist.c
@@ -108,10 +108,10 @@ static int SubstitutePhonemes(PHONEME_LIST *plist_out, PHONEME_LIST2 *ph_list2, 
 	return n_plist_out;
 }
 
-void MakePhonemeList(const Translator *tr, PHONEME_LIST2 *ph_list2, int *n_ph_list2p, int post_pause, bool start_sentence)
+void MakePhonemeList(const Translator *tr, PHONEME_LIST2 *ph_list2, uint16_t *n_ph_list2p, int post_pause, bool start_sentence)
 {
 	int ix = 0;
-        int n_ph_list2_cur = *n_ph_list2p;
+		int n_ph_list2_cur = *n_ph_list2p;
 	int j;
 	int insert_ph = 0;
 	PHONEME_LIST *phlist;
@@ -186,10 +186,10 @@ void MakePhonemeList(const Translator *tr, PHONEME_LIST2 *ph_list2, int *n_ph_li
 
 		if (plist2[j].phcode == phonSWITCH) {
 			if ((!(plist2[j].synthflags & SFLAG_EMBEDDED)) && (
-			        (plist2[j].tone_ph == current_phoneme_tab) ||
-			        (plist2[j+1].phcode == phonSWITCH) ||
-			        ((plist2[j+1].phcode == phonPAUSE) && (plist2[j+2].phcode == phonSWITCH))
-			        )) {
+					(plist2[j].tone_ph == current_phoneme_tab) ||
+					(plist2[j+1].phcode == phonSWITCH) ||
+					((plist2[j+1].phcode == phonPAUSE) && (plist2[j+2].phcode == phonSWITCH))
+					)) {
 				// delete this phonSWITCH if it's switching to the current phoneme table, or
 				// delete this phonSWITCH if its followed by another phonSWITCH
 				if (deleted_sourceix == -1 && plist2[j].sourceix != 0)
